requirements.txt
# ── Core scraping ─────────────────────────────────────────────────────────────
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.40.0
playwright-stealth>=1.0.0
pyyaml>=6.0

# ── Data processing & export ──────────────────────────────────────────────────
pandas>=2.0.0
tqdm>=4.66.0
openpyxl>=3.1.0
brotli>=1.1.0

# ── Database ──────────────────────────────────────────────────────────────────
asyncpg>=0.29.0
aiomysql>=0.2.0
aiosqlite>=0.19.0

# ── API platform ──────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0

# ── AI / LLM integrations ────────────────────────────────────────────────────
openai>=1.0.0

# ── Notifications ─────────────────────────────────────────────────────────────
slack-sdk>=3.0.0

# ── NLP enrichment layer ─────────────────────────────────────────────────────
langdetect>=1.0.9
yake>=0.4.8
sumy>=0.9.0
nltk>=3.8.1

# ── Environment / config ─────────────────────────────────────────────────────
python-dotenv>=1.0.1
structlog>=24.0.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
